add: format function
